Continue medical insurance while on leave
I feel we're getting a better handle on leaves of absence now. They are no
longer open-ended, and the people going on leave fully plan to return. While
they are on-leave, they are often traveling, doing some marketing for us, or are
generally representing thoughtbot.

We do not want to create a burden on people who have requested a leave for
personal reasons, are not collecting salary, but are still a member of our team
and will be resuming work in the near future.

This change intends to make it so that we continue medical insurance while
someone is on unpaid leave.