Apprentices now receive the same benefits
Over the last 12 months there has only been one apprentice (out of 13) whom we
did not promote to a full designer or developer at the completion of the
apprenticeship.

We have done an excellent job of setting the bar very high for getting into the
apprenticeship, and training people in it.

The majority (62%) of apprentices in the last 12 months were already working as
a designer or developer, and left (or recently left) their job to join us as an
apprentice.

We offer a reduced salary during apprenticeship, but don't want to create an
additional burden on apprentices.

Also, by negotiating medical benefits for some apprentices and not others, we
run the risk of our medical plan being viewed as discriminatory, which we
certainly want to avoid.

This change make apprentices eligible, as a full-time employee of thoughtbot,
for the same benefits as all other full-time employees.