Remove official documentation on access to files
I'm hoping to removing the additional, unecessary stuff from the handbook that
is left over from the official handbook we had created by paychex years ago.

This particular policy is unecessary. There are no federal laws about being able
to review your file, but there are laws in many states about this.

If someone wanted to view the records, we would let them, but I don't think we
need this corporate-speak policy in the handbook.